+++ instrumentation/java-http-server/library/src/main/java/io/opentelemetry/instrumentation/javahttpserver/JavaHttpServerTelemetryBuilder.java
@@ -64,14 +64,14 @@ public final class JavaHttpServerTelemetryBuilder {
   /**
    * Configures which HTTP request headers are captured as span attributes.
    *
-   * <p>Header values are captured under the {@code http.request.header.<key>} attribute key, where
-   * {@code <key>} is the lowercase header name.
+   * <p>Header values are captured under the {@code http.request.header.<key>} attribute key. The
+   * {@code <key>} part in the attribute key is the lowercase header name.
    *
    * <p>Selector patterns are matched case-insensitively, since HTTP header names are
    * case-insensitive. {@code ?} matches one character and {@code *} matches any number of
    * characters, including none. Excluded patterns take precedence over included patterns. A
-   * selector with no included patterns captures every header that is not excluded, while an
-   * {@linkplain IncludeExclude#isEmpty() empty} selector, like an absent one, captures no headers.
+   * selector with no included patterns captures every header that is not excluded, and an
+   * {@linkplain IncludeExclude#isEmpty() empty} selector captures no headers.
    */
   @CanIgnoreReturnValue
   public JavaHttpServerTelemetryBuilder setRequestHeaders(IncludeExclude requestHeaders) {
@@ -82,9 +82,8 @@ public final class JavaHttpServerTelemetryBuilder {
   /**
    * Configures HTTP request headers to capture as span attributes.
    *
-   * <p>Header names not containing {@code *} or {@code ?} are matched literally. Header names
-   * containing either character are ignored and logged, since this setting never supported them as
-   * wildcards. Use {@link #setRequestHeaders(IncludeExclude)} to match names by pattern.
+   * <p>The header names are matched literally, so {@code *} and {@code ?} are not treated as glob
+   * patterns.
    *
    * @param requestHeaders HTTP header names to capture.
    * @deprecated Use {@link #setRequestHeaders(IncludeExclude)} instead, which matches glob patterns
@@ -101,14 +100,14 @@ public final class JavaHttpServerTelemetryBuilder {
   /**
    * Configures which HTTP response headers are captured as span attributes.
    *
-   * <p>Header values are captured under the {@code http.response.header.<key>} attribute key, where
-   * {@code <key>} is the lowercase header name.
+   * <p>Header values are captured under the {@code http.response.header.<key>} attribute key. The
+   * {@code <key>} part in the attribute key is the lowercase header name.
    *
    * <p>Selector patterns are matched case-insensitively, since HTTP header names are
    * case-insensitive. {@code ?} matches one character and {@code *} matches any number of
    * characters, including none. Excluded patterns take precedence over included patterns. A
-   * selector with no included patterns captures every header that is not excluded, while an
-   * {@linkplain IncludeExclude#isEmpty() empty} selector, like an absent one, captures no headers.
+   * selector with no included patterns captures every header that is not excluded, and an
+   * {@linkplain IncludeExclude#isEmpty() empty} selector captures no headers.
    */
   @CanIgnoreReturnValue
   public JavaHttpServerTelemetryBuilder setResponseHeaders(IncludeExclude responseHeaders) {
@@ -119,9 +118,8 @@ public final class JavaHttpServerTelemetryBuilder {
   /**
    * Configures HTTP response headers to capture as span attributes.
    *
-   * <p>Header names not containing {@code *} or {@code ?} are matched literally. Header names
-   * containing either character are ignored and logged, since this setting never supported them as
-   * wildcards. Use {@link #setResponseHeaders(IncludeExclude)} to match names by pattern.
+   * <p>The header names are matched literally, so {@code *} and {@code ?} are not treated as glob
+   * patterns.
    *
    * @param responseHeaders HTTP header names to capture.
    * @deprecated Use {@link #setResponseHeaders(IncludeExclude)} instead, which matches glob
